feat(combobox): update stories
