+++ core/components/organisms/combobox/__stories__/size.story.jsx
@@ -116,7 +116,7 @@ const customCode = `() => {
 }`;
 
 export default {
-  title: 'Components/Combobox/Single Select/Sizes',
+  title: 'Components/Combobox/Sizes',
   component: Combobox,
   parameters: {
     docs: {
