--- css/src/components/combobox.css
@@ -1,11 +1,0 @@
-.Combobox-Option--focus {
-  /* outline: 3px auto var(--secondary-shadow); */
-  /* background-color: var(--secondary-lightest); */
-  background-color: red;
-}
-
-.myclass:focus,
-.myclass:focus-visible {
-  background-color: red;
-  outline: 3px solid green;
-}
