+++ css/src/components/listbox.css
@@ -47,8 +47,7 @@
 
 .Listbox-item--option:focus,
 .Listbox-item--option:focus-visible {
-  /* outline: 3px auto var(--secondary-shadow); */
-  outline: 3px auto green;
+  outline: 3px auto var(--secondary-shadow);
 }
 
 .Listbox-item--option:active {
